README.md
Isn't it so pretty and lovely?
This is a lovely website from General Assembly. As I annoyingly repeat myself again, this was the example used for their Intro to Coding webinar. I am unsure if it's on their own GitHub, but if it is, I will take the correct action to either credit them or fork the repo directly from them. I have a horrid memory and wanted to get this saved asap, so I forked it. I forked it up. This is public, however, hence the really long-winded and possibly unnecessary disclaimer.